DFS on Binary Search Tree
Here we have two base cases, when the value is not present in the
tree, ie: null, or when we find the needle.

If neither of these are true then we continue to recurse.

The runtime of this algorithm varies according to the height of
the tree and how balanced the tree is.
Worst case: O(n)
Best case: O(log n)