Heap
A heap is a data structure which lends itself well to a priority
queue.

A heap is self balancing as adding/deleting means we adjust the tree.
A heap is usually a complete/balanced tree.

Since a heap is usually a balanced tree the running time of the
insert/delete methods are O(log n).